+++ Components/Volume/src/OgreVolumeGridSource.cpp
@@ -235,12 +235,12 @@ namespace Volume {
         float value;
         int x, y;
         Vector3 scaledCenter(center.x * mPosXScale, center.y * mPosYScale, center.z * mPosZScale);
-        int xStart = Math::Clamp((size_t)(scaledCenter.x - radius * mPosXScale), (size_t)0, mWidth);
-        int xEnd = Math::Clamp((size_t)(scaledCenter.x + radius * mPosXScale), (size_t)0, mWidth);
-        int yStart = Math::Clamp((size_t)(scaledCenter.y - radius * mPosYScale), (size_t)0, mHeight);
-        int yEnd = Math::Clamp((size_t)(scaledCenter.y + radius * mPosYScale), (size_t)0, mHeight);
-        int zStart = Math::Clamp((size_t)(scaledCenter.z - radius * mPosZScale), (size_t)0, mDepth);
-        int zEnd = Math::Clamp((size_t)(scaledCenter.z + radius * mPosZScale), (size_t)0, mDepth);
+        int xStart = Math::Clamp(static_cast<int>(scaledCenter.x - radius * mPosXScale), 0, static_cast<int>(mWidth));
+        int xEnd = Math::Clamp(static_cast<int>(scaledCenter.x + radius * mPosXScale), 0, static_cast<int>(mWidth));
+        int yStart = Math::Clamp(static_cast<int>(scaledCenter.y - radius * mPosYScale), 0, static_cast<int>(mHeight));
+        int yEnd = Math::Clamp(static_cast<int>(scaledCenter.y + radius * mPosYScale), 0, static_cast<int>(mHeight));
+        int zStart = Math::Clamp(static_cast<int>(scaledCenter.z - radius * mPosZScale), 0, static_cast<int>(mDepth));
+        int zEnd = Math::Clamp(static_cast<int>(scaledCenter.z + radius * mPosZScale), 0, static_cast<int>(mDepth));
         Vector3 pos;
         for (int z = zStart; z < zEnd; ++z)
         {
@@ -249,7 +249,7 @@ namespace Volume {
                 for (x = xStart; x < xEnd; ++x)
                 {
                     pos.x = x * worldWidthScale;
-                    pos.y =  y * worldHeightScale;
+                    pos.y = y * worldHeightScale;
                     pos.z = z * worldDepthScale;
                     value = operation->getValue(pos);
                     setVolumeGridValue(x, y, z, value);
