+++ Components/Volume/src/OgreVolumeOctreeNode.cpp
@@ -34,7 +34,7 @@ namespace Volume {
     
     const Real OctreeNode::NEAR_FACTOR = (Real)2.0;
     const size_t OctreeNode::OCTREE_CHILDREN_COUNT = 8;
-    size_t OctreeNode::mGridPositionCount = 0;
+    uint32 OctreeNode::mGridPositionCount = 0;
     size_t OctreeNode::mNodeI = 0;
     
     //-----------------------------------------------------------------------
